fix vitrikushIhor types
