+++ my-app/src/common/appTypes.ts
@@ -1,6 +1,20 @@
 import { actionsPosts } from "../redux/actions/actionsPosts";
+
 import { AppStateType, PropertiesTypes } from "../redux/store";
 import { ThunkAction } from "redux-thunk";
+
+export enum actionTypes {
+  SET_POSTS = "SET_POSTS",
+  SET_IS_LOADING = "SET_IS_LOADING",
+  Vitrikush_SET_LOGIN = "Vitrikush_SET_LOGIN"
+}
+
+export interface TypeLogin {
+  email: string,
+  name: string,
+  password: string,
+  verifyPassword: string,
+}
 
 export interface Post {
   title: string;
