+++ my-app/src/redux/store.ts
@@ -6,7 +6,6 @@ import { composeWithDevTools } from "redux-devtools-extension/developmentOnly";
 
 const rootReducer = combineReducers({
   postsReducer: postsReducer,
-  // @ts-ignore
   VitrikushReducer,
 })
